// types/transit.ts
export type TransitMode =
  | 'subway'
  | 'bus'
  | 'lirr'
  | 'metroNorth'
  | 'accessARide'
  | 'bridgesTunnels'
  | 'sir'
// ...
export interface RecoveryData {
  date: string // ISO date string
  subwayPct: number
  busPct: number
  lirrPct: number
  metroNorthPct: number
  accessARidePct: number
  bridgesTunnelsPct: number
  sirPct: number
}